Honor language specific settings.
When formatting an entire file, use the file's scope when determining
what values to use for softTabs and tabLength. When formatting selected
text within a file, default to the `source.json` scope if available.
In either case, always fallback to the editor defaults when language
specific settings are not available.